refactor wave operations handoff ai router
